Refactor Dynamics to MVVM
